+++ slim/base/log.py
@@ -28,7 +28,8 @@ def enable(level=logging.DEBUG):
     if not is_enable:
         default_handler = logging.StreamHandler()
         default_handler.setFormatter(logging.Formatter(
-            '[%(asctime)s][%(levelname).4s][slim] %(message)s'
+            # '[%(asctime)s][%(levelname).4s][slim] %(message)s'
+            '[%(levelname).4s] %(message)s'
         ))
 
         logger = logging.getLogger('slim')