+++ slim/base/route.py
@@ -3,20 +3,23 @@ from asyncio import iscoroutinefunction, Future
 from typing import Iterable, Type, TYPE_CHECKING
 from aiohttp import web, web_response
 from posixpath import join as urljoin
+
+from aiohttp.abc import Request
 from slim.base.ws import WSRouter
 from ..utils.async_run import sync_call, async_call
 
 if TYPE_CHECKING:
     from .view import BaseView
+    from .app import Application
 
 logger = logging.getLogger(__name__)
 # __all__ = ('Route',)
 
 
-def get_route_middleware(app):
+def get_route_middleware(app: 'Application'):
     @web.middleware
     # noinspection PyProtectedMember
-    async def route_middleware(request, handler):
+    async def route_middleware(request: Request, handler):
         if not app.route._is_beacon(handler):
             return await handler(request)
         else:
@@ -24,37 +27,42 @@ def get_route_middleware(app):
             view_cls = data['view']
             func = data['handler']
 
-            handler_name = '%s.%s' % (view_cls.__name__, func.__name__)
-            ascii_encodable_path = request.path.encode('ascii', 'backslashreplace').decode('ascii')
+            handler_name = '%s.%s.%s' % (view_cls.__module__, view_cls.__name__, func.__name__)
+            ascii_encodable_path = request.path_qs.encode('ascii', 'backslashreplace').decode('ascii')
+            status_code = 200
 
-            # GET /api/get -> TopicView.get
-            logger.info("{} {} -> {}".format(request._method, ascii_encodable_path, handler_name))
             view_instance: BaseView = view_cls(app, request)
 
             from .view import ErrorCatchContext
 
             with ErrorCatchContext(view_instance):
                 await view_instance._prepare()
-            if view_instance.is_finished:
-                return view_instance.response
 
-            with ErrorCatchContext(view_instance):
-                await async_call(view_instance.prepare)
             if view_instance.is_finished:
-                return view_instance.response
+                resp = view_instance.response
+            else:
+                resp = await func(view_instance) or view_instance.response
+                if not isinstance(resp, web_response.StreamResponse):
+                    status_code = 500
 
-            resp = await func(view_instance) or view_instance.response
-            assert isinstance(resp, web_response.StreamResponse), \
-                "Handler {!r} should return response instance, got {!r}. Is view.finish() not be called?".format(handler_name, type(resp), )
+            # GET /api/get -> TopicView.get 200
+            logger.info("{} {:4s} -> {} {}".format(request._method, ascii_encodable_path, handler_name, status_code))
+
+            if logger.isEnabledFor(logging.DEBUG):
+                logger.debug('query parameters: %s', view_instance.params)
+                logger.debug('post data: %s', await view_instance.post_data())
+
+            if status_code == 500:
+                resp.set_status(status_code, "The handler {!r} did not called `view.finish()`.".format(handler_name))
+                return resp
 
             await view_instance._on_finish()
-            await async_call(view_instance.on_finish)
             return resp
 
     return route_middleware
 
 
-def view_bind(app, cls_url, view_cls: Type['BaseView']):
+def view_bind(app: 'Application', cls_url, view_cls: Type['BaseView']):
     """
     将 API 绑定到 web 服务上
     :param view_cls:
@@ -65,30 +73,41 @@ def view_bind(app, cls_url, view_cls: Type['BaseView']):
     if view_cls._no_route: return
     cls_url = cls_url or view_cls.__class__.__name__.lower()
 
-    def add_route(name, route_info, beacon_info):
-        for method in route_info['method']:
+    def add_route(beacon_info):
+        route = beacon_info['route']
+        for method in route['method']:
             async def beacon(request): pass
-            route_key = route_info['url'] if route_info['url'] else name
-            app._raw_app.router.add_route(method, urljoin('/api', cls_url, route_key), beacon)
+            app._raw_app.router.add_route(method, route['fullpath'], beacon)
             app.route._beacons[beacon] = beacon_info
 
     # noinspection PyProtectedMember
     for name, route_info_lst in view_cls._interface.items():
         for route_info in route_info_lst:
             real_handler = getattr(view_cls, name, None)
-            if real_handler is None: continue # TODO: delete
+            if real_handler is None: continue  # TODO: delete
             assert real_handler is not None, "handler must be exists"
 
-            handler_name = '%s.%s' % (view_cls.__name__, real_handler.__name__)
-            assert iscoroutinefunction(real_handler), "Add 'async' before interface function %r" % handler_name
+            handler_name = '%s.%s.%s' % (view_cls.__module__, view_cls.__name__, real_handler.__name__)
+            if not iscoroutinefunction(real_handler):
+                logger.error("Interface function must be async: %r" % handler_name)
+                exit(-1)
+
+            cls_url = cls_url or view_cls.__class__.__name__.lower()
+            route_key = route_info['url'] or name
 
             beacon_info = {
                 'view': view_cls,
-                'name': name,
+                'name': name,  # name of function
                 'handler': real_handler,
-                'route_info': route_info
+                'handler_name': handler_name,
+                'route': {
+                    'method': route_info['method'],  # Set[HttpMethod]
+                    'relpath': route_key,
+                    'fullpath': urljoin(app.mountpoint, cls_url, route_key)
+                }
             }
-            add_route(name, route_info, beacon_info)
+
+            add_route(beacon_info)
 
 
 class Route: